Several additions: * Rate control for recovery added in WebRTCConn * bandwidth-alert is now a StreamEvent * Modified detection and recovery algorithm for BW in erizoJSController * Modified basic_example to show the new event

+++ erizo/src/erizo/SdpInfo.cpp
@@ -328,7 +328,7 @@ namespace erizo {
       if (bundleTags.size()>2){
         ELOG_WARN("More bundleTags than supported, expect unexpected behaviour");
       }
-      for (int i = 0; i < bundleTags.size(); i++){
+      for (uint8_t i = 0; i < bundleTags.size(); i++){
         if(bundleTags[i].mediaType == AUDIO_TYPE){
           sdp << "a=mid:" << bundleTags[i].id << endl;
         }
@@ -422,7 +422,7 @@ namespace erizo {
           sdp << "a=recvonly" << endl;
           break;
       }
-      for (int i = 0; i < bundleTags.size(); i++){
+      for (uint8_t i = 0; i < bundleTags.size(); i++){
         if(bundleTags[i].mediaType == VIDEO_TYPE){
           sdp << "a=mid:" << bundleTags[i].id << endl;
         }
